Organize the package
