+++ .idea/misc.xml
@@ -12,7 +12,7 @@
         <entry key="..\:/Users/LENOVO/AndroidStudioProjects/ProgrammingLanguage/app/src/main/res/layout/activity_about.xml" value="0.264" />
         <entry key="..\:/Users/LENOVO/AndroidStudioProjects/ProgrammingLanguage/app/src/main/res/layout/activity_detail.xml" value="0.14583333333333334" />
         <entry key="..\:/Users/LENOVO/AndroidStudioProjects/ProgrammingLanguage/app/src/main/res/layout/activity_main.xml" value="0.14583333333333334" />
-        <entry key="..\:/Users/LENOVO/AndroidStudioProjects/ProgrammingLanguage/app/src/main/res/layout/item_cardview_language.xml" value="0.14583333333333334" />
+        <entry key="..\:/Users/LENOVO/AndroidStudioProjects/ProgrammingLanguage/app/src/main/res/layout/item_cardview_language.xml" value="0.4983922829581994" />
         <entry key="..\:/Users/LENOVO/AndroidStudioProjects/ProgrammingLanguage/app/src/main/res/layout/item_grid_language.xml" value="0.14583333333333334" />
         <entry key="..\:/Users/LENOVO/AndroidStudioProjects/ProgrammingLanguage/app/src/main/res/layout/item_row_language.xml" value="0.14583333333333334" />
         <entry key="..\:/Users/LENOVO/AndroidStudioProjects/ProgrammingLanguage/app/src/main/res/menu-v26/menu_detail.xml" value="0.27291666666666664" />
